Break out the http::RequestManager from HTTPClient.h
