+++ Common/Common.vcxproj
@@ -453,6 +453,7 @@
     <ClInclude Include="Math\Statistics.h" />
     <ClInclude Include="Net\Connection.h" />
     <ClInclude Include="Net\HTTPNaettRequest.h" />
+    <ClInclude Include="Net\HTTPRequestManager.h" />
     <ClInclude Include="Net\NetBuffer.h" />
     <ClInclude Include="Net\HTTPClient.h" />
     <ClInclude Include="Net\HTTPHeaders.h" />
@@ -909,6 +910,7 @@
     <ClCompile Include="Math\Statistics.cpp" />
     <ClCompile Include="Net\Connection.cpp" />
     <ClCompile Include="Net\HTTPNaettRequest.cpp" />
+    <ClCompile Include="Net\HTTPRequestManager.cpp" />
     <ClCompile Include="Net\NetBuffer.cpp" />
     <ClCompile Include="Net\HTTPClient.cpp" />
     <ClCompile Include="Net\HTTPHeaders.cpp" />
