+++ Common/Common.vcxproj.filters
@@ -692,6 +692,9 @@
     <ClInclude Include="Net\Connection.h">
       <Filter>Net</Filter>
     </ClInclude>
+    <ClInclude Include="Net\HTTPRequestManager.h">
+      <Filter>Net</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="ABI.cpp" />
@@ -1294,6 +1297,9 @@
     </ClCompile>
     <ClCompile Include="LoongArch64Emitter.cpp" />
     <ClCompile Include="Net\Connection.cpp">
+      <Filter>Net</Filter>
+    </ClCompile>
+    <ClCompile Include="Net\HTTPRequestManager.cpp">
       <Filter>Net</Filter>
     </ClCompile>
   </ItemGroup>
