+++ Common/Net/HTTPClient.h
@@ -92,20 +92,4 @@ private:
 	net::ResolveFunc customResolve_;
 };
 
-// Fake request for cache hits.
-// The download manager uses this when caching was requested, and a new-enough file was present in the cache directory.
-// This is simply a finished request, that can still be queried like a normal one so users don't know it came from the cache.
-class CachedRequest : public Request {
-public:
-	CachedRequest(RequestMethod method, std::string_view url, std::string_view name, bool *cancelled, RequestFlags flags, std::string_view responseData)
-		: Request(method, url, name, cancelled, flags)
-	{
-		buffer_.Append(responseData);
-	}
-	void Start() override {}
-	void Join() override {}
-	bool Done() override { return true; }
-	bool Failed() const override { return false; }
-};
-
 }  // namespace http
