+++ Common/Net/HTTPRequest.h
@@ -96,57 +96,6 @@ private:
 	std::function<void(Request &)> callback_;
 };
 
-using std::shared_ptr;
-
-class RequestManager {
-public:
-	~RequestManager() {
-		CancelAll();
-	}
-
-	// NOTE: This is the only version that supports the cache flag (for now).
-	std::shared_ptr<Request> StartDownload(std::string_view url, const Path &outfile, RequestFlags flags, const char *acceptMime = nullptr);
-
-	std::shared_ptr<Request> StartDownloadWithCallback(
-		std::string_view url,
-		const Path &outfile,
-		RequestFlags flags,
-		std::function<void(Request &)> callback,
-		std::string_view name = "",
-		const char *acceptMime = nullptr);
-
-	std::shared_ptr<Request> AsyncPostWithCallback(
-		std::string_view url,
-		std::string_view postData,
-		std::string_view postMime, // Use postMime = "application/x-www-form-urlencoded" for standard form-style posts, such as used by retroachievements. For encoding form data manually we have MultipartFormDataEncoder.
-		RequestFlags flags,
-		std::function<void(Request &)> callback,
-		std::string_view name = "");
-
-	// Drops finished downloads from the list.
-	void Update();
-	void CancelAll();
-
-	void SetUserAgent(std::string_view userAgent) {
-		userAgent_ = userAgent;
-	}
-
-	void SetCacheDir(const Path &path) {
-		cacheDir_ = path;
-	}
-
-	Path UrlToCachePath(const std::string_view url);
-
-private:
-	std::vector<std::shared_ptr<Request>> downloads_;
-	// These get copied to downloads_ in Update(). It's so that callbacks can add new downloads
-	// while running.
-	std::vector<std::shared_ptr<Request>> newDownloads_;
-
-	std::string userAgent_;
-	Path cacheDir_;
-};
-
 inline const char *RequestMethodToString(RequestMethod method) {
 	switch (method) {
 	case RequestMethod::GET: return "GET";
