+++ Core/Config.cpp
@@ -41,11 +41,11 @@
 #include "Common/File/VFS/VFS.h"
 #include "Common/Log/LogManager.h"
 #include "Common/Math/CrossSIMD.h"
+#include "Common/Net/HTTPRequestManager.h"
 #include "Common/OSVersion.h"
 #include "Common/System/Display.h"
 #include "Common/System/System.h"
 #include "Common/StringUtils.h"
-#include "Common/Thread/ThreadUtil.h"
 #include "Common/GPU/Vulkan/VulkanLoader.h"
 #include "Common/VR/PPSSPPVR.h"
 #include "Common/System/OSD.h"
