+++ Core/HLE/sceNet.cpp
@@ -25,6 +25,7 @@
 #include "Common/Data/Text/I18n.h"
 #include "Common/File/VFS/VFS.h"
 #include "Common/File/FileUtil.h"
+#include "Common/Net/HTTPRequestManager.h"
 #include "Common/System/OSD.h"
 #include "Common/Serialize/Serializer.h"
 #include "Common/Serialize/SerializeFuncs.h"
