+++ Core/Reporting.cpp
@@ -40,6 +40,7 @@ extern "C" {
 #include "Common/System/OSD.h"
 #include "Common/Data/Text/I18n.h"
 #include "Common/Net/HTTPClient.h"
+#include "Common/Net/HTTPRequestManager.h"
 #include "Common/Net/Resolve.h"
 #include "Common/Net/URL.h"
 #include "Common/Thread/ThreadUtil.h"
