+++ Core/RetroAchievements.cpp
@@ -31,6 +31,7 @@
 #include "Common/Log.h"
 #include "Common/File/Path.h"
 #include "Common/Net/HTTPRequest.h"
+#include "Common/Net/HTTPRequestManager.h"
 #include "Common/System/OSD.h"
 #include "Common/System/System.h"
 #include "Common/System/NativeApp.h"
