+++ UI/DevScreens.cpp
@@ -39,7 +39,7 @@
 #include "Common/Data/Text/I18n.h"
 #include "Common/Data/Text/Parsers.h"
 #include "Common/Data/Encoding/Utf8.h"
-#include "Common/Net/HTTPClient.h"
+#include "Common/Net/HTTPRequestManager.h"
 #include "Common/UI/Context.h"
 #include "Common/UI/View.h"
 #include "Common/UI/ViewGroup.h"
