+++ UWP/CommonUWP/CommonUWP.vcxproj
@@ -102,6 +102,7 @@
     <ClInclude Include="..\..\Common\Net\Connection.h" />
     <ClInclude Include="..\..\Common\Net\HTTPNaettRequest.h" />
     <ClInclude Include="..\..\Common\Net\HTTPRequest.h" />
+    <ClInclude Include="..\..\Common\Net\HTTPRequestManager.h" />
     <ClInclude Include="..\..\Common\Net\NetBuffer.h" />
     <ClInclude Include="..\..\Common\Data\Collections\ConstMap.h" />
     <ClInclude Include="..\..\Common\Data\Collections\FixedSizeQueue.h" />
@@ -289,6 +290,7 @@
     <ClCompile Include="..\..\Common\Net\Connection.cpp" />
     <ClCompile Include="..\..\Common\Net\HTTPNaettRequest.cpp" />
     <ClCompile Include="..\..\Common\Net\HTTPRequest.cpp" />
+    <ClCompile Include="..\..\Common\Net\HTTPRequestManager.cpp" />
     <ClCompile Include="..\..\Common\Net\NetBuffer.cpp" />
     <ClCompile Include="..\..\Common\Data\Color\RGBAUtil.cpp" />
     <ClCompile Include="..\..\Common\Data\Convert\SmallDataConvert.cpp" />
