+++ libretro/Makefile.common
@@ -447,6 +447,7 @@ SOURCES_CXX += \
 	$(COMMONDIR)/Net/HTTPServer.cpp \
 	$(COMMONDIR)/Net/HTTPRequest.cpp \
 	$(COMMONDIR)/Net/HTTPNaettRequest.cpp \
+	$(COMMONDIR)/Net/HTTPRequestManager.cpp \
 	$(COMMONDIR)/Net/NetBuffer.cpp \
 	$(COMMONDIR)/Net/Resolve.cpp \
 	$(COMMONDIR)/Net/Sinks.cpp \
